// src/app/api/model/clientSummaryDto.ts
/**
 * GestionRFQ API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ClientSummaryDto { 
    id?: number;
    nom?: string | null;
    email?: string | null;
    sales?: string | null;
}