// src/app/api/model/generateReportRequest.ts
/**
 * GestionRFQ API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GenerateReportRequest { 
    startDate?: string | null;
    endDate?: string | null;
    marketSegmentIds?: Array<number> | null;
    clientIds?: Array<number> | null;
    statuts?: Array<string> | null;
    format?: string | null;
}